module-4-rag-evaluation/rag-system/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
anthropic==0.18.0
openai==1.12.0
httpx<0.28
google-generativeai==0.3.0
python-dotenv==1.0.0
# Optional vector backend is skipped on Python 3.12+ to avoid
# chroma-hnswlib native build failures on Windows.
chromadb==0.4.22; python_version < "3.12"
sentence-transformers==2.3.0; python_version < "3.12"
